requirements.txt
# --- Core Django & API ---
Django==4.2.20
djangorestframework==3.15.2
django-cors-headers==4.4.0



# --- Deployment (Render) ---
gunicorn==23.0.0
whitenoise==6.8.2
dj-database-url==2.3.0
python-dotenv==1.0.1

# --- Django internal dependencies ---
asgiref==3.8.1
sqlparse==0.5.1
tzdata==2024.2

mysqlclient==2.2.0
